pyproject.toml
[project]
name = "a2a-mcp-langgraph-agent-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.3.7",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiomysql==0.2.0",
    "aiosignal==1.4.0",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "arxiv>=2.2.0",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "brotli==1.1.0",
    "cachetools==6.2.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cryptography==46.0.2",
    "dataclasses-json==0.6.7",
    "ddgs==9.6.0",
    "duckduckgo-search>=8.1.1",
    "fastapi==0.118.0",
    "frozenlist==1.8.0",
    "google-api-core==2.25.2",
    "google-auth==2.41.1",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.75.1",
    "grpcio-reflection==1.75.1",
    "grpcio-tools==1.75.1",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.27",
    "langchain-community==0.3.30",
    "langchain-core==0.3.77",
    "langchain-mcp-adapters",
    "langchain-ollama==0.3.10",
    "langchain-text-splitters==0.3.11",
    "langgraph==0.6.8",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.32",
    "lxml==6.0.2",
    "marshmallow==3.26.1",
    "mcp",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "numpy==2.3.3",
    "ollama==0.6.0",
    "opentelemetry-api==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-semantic-conventions==0.58b0",
    "orjson==3.11.3",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "primp==0.15.0",
    "propcache>=0.2.1,!=0.4.0",
    "proto-plus==1.26.1",
    "protobuf==6.32.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.11.0",
    "pymysql==1.1.2",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.48.0",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.37.0",
    "wikipedia>=1.4.0",
    "xxhash==3.6.0",
    "yarl==1.21.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
